requirements.txt
python-barcode
pywin32
pillow
pyqt5